Add files via upload
Description: This project utilizes multinomial distributions to perform Text Clustering on a Bag of Words model. 

Input: To keep the model simple I kept the collection unordered, disregarding grammar and word order. Then, I count how often each word appears in each document and store the word counts in a matrix, where each row of the matrix represents one document. 

Output: Learned conditional distribution matrix which can be used to find the highest frequency topics in the text file

Procedure: In order to do text clustering, I used a mixture of multinomial distributions so that each topic has a particular multinomial distribution associated with it. As a result, each document is a mixture of different topics. In order to cluster a set of documents, I fit this mixture model to the data. In this problem, the EM algorithm is used for fitting mixture models.